Add next-auth, session provider, and useSession, setup some testing logic
The .env file in this is not for production, it is only for local mode. There are some commented out sections that will need to be addressed.

Still need to add logic for actually handling login event. Also need to add credentials provider for manually logging in (email/pass).

OAuth is ready for when the component is made and backend has clientId and secrets ready.